requirements.txt
python-dotenv
python-wordpress-xmlrpc
